moved main code to runner package under epp/cmd
